+++ utils_ver.py
@@ -5,7 +5,7 @@ import time
 
 import requests
 
-__version__ = "20221120-4-010"
+__version__ = "20231216-1-001"
 ver_re = re.compile("__version__ = .(\\d+-\\d+-...).")
 
 
@@ -15,8 +15,8 @@ def get_present_ver() -> str:
 
 def get_latest_ver() -> str:
     url = "https://ghproxy.com/https://raw.githubusercontent.com/yczcc/checkinpanel/master/utils_ver.py"
-    if time.localtime().tm_hour < 8 or time.localtime().tm_hour > 12:
-        return "不在 8-12 点内，跳过版本检查。"
+    if time.localtime().tm_hour < 8 or time.localtime().tm_hour > 18:
+        return "不在 8-18 点内，跳过版本检查。"
     try:
         r = requests.get(url=url, timeout=3)
     except Exception as e: